Inline OIDC request schema in the view file
